requirements.txt
-e git+https://github.com/barseghyanartur/django-qartez@master#egg=django-qartez
-e git+https://github.com/dchaplinsky/grappelli-modeltranslation.git#egg=master
cairocffi==0.8.0
CairoSVG==1.0.16
cffi==1.11.5
cryptography==2.3.1
django-cacheops==4.0.1
django-grappelli==2.10.1
django-jinja==2.3.1
django-markdown==0.8.4
django-modeltranslation==0.12.1
django-nested-admin==3.0.21
django-pickling==1.0
django-pipeline==1.6.13
django-recaptcha==1.0.4
django-select2-forms==2.0.2
django-wysiwyg-redactor==0.5.1
Django==1.11.23
easy-thumbnails==2.4.1
elasticsearch-dsl==5.3.0
fuzzywuzzy==0.16.0
gspread==0.2.5
Jinja2==2.10.1
jsonfield==1.0.3
oauth2client==1.4.12
psycopg2
pyOpenSSL==17.5.0
python-levenshtein==0.12.0
raven==6.3.0
requests>=2.20.0
rfc6266==0.0.4
Pillow>=6.2.0
tqdm==4.19.8
translitua==1.3.0
unicodecsv==0.9.4
wagtail==1.13.4
WeasyPrint==0.24
XlsxWriter==1.0.0
django-compressor==2.2
django-otp==0.4.3
qrcode==6.0
jmespath==0.9.3
django-neomodel==0.0.4